Add Microsoft.DotNet.Interactive.AspNetCore project
