+++ dotnet-interactive.sln
@@ -74,6 +74,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{4DD50BCA-1
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Interactive.Http", "src\Microsoft.DotNet.Interactive.Http\Microsoft.DotNet.Interactive.Http.csproj", "{DE942DC3-B41B-45C5-9E01-3C4F042D3916}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Interactive.AspNetCore", "src\Microsoft.DotNet.Interactive.AspNetCore\Microsoft.DotNet.Interactive.AspNetCore.csproj", "{5167FC61-2FBF-4E0D-877E-C2DD354FB955}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -360,6 +362,18 @@ Global
 		{DE942DC3-B41B-45C5-9E01-3C4F042D3916}.Release|x64.Build.0 = Release|Any CPU
 		{DE942DC3-B41B-45C5-9E01-3C4F042D3916}.Release|x86.ActiveCfg = Release|Any CPU
 		{DE942DC3-B41B-45C5-9E01-3C4F042D3916}.Release|x86.Build.0 = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|x64.Build.0 = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Debug|x86.Build.0 = Debug|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|x64.ActiveCfg = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|x64.Build.0 = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|x86.ActiveCfg = Release|Any CPU
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -388,6 +402,7 @@ Global
 		{8017943B-B591-49C8-B957-40141B427AD9} = {979A5022-09CA-4258-B76D-CFDC31DA71E0}
 		{9A091E80-D5EB-4DEF-8E38-B24B128393DF} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 		{DE942DC3-B41B-45C5-9E01-3C4F042D3916} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
+		{5167FC61-2FBF-4E0D-877E-C2DD354FB955} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6D05A9AF-CFFB-4187-8599-574387B76727}
