+++ src/Microsoft.DotNet.Interactive.AspNetCore/EndpointRouteBuilderExtensions.cs
@@ -5,19 +5,19 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 
-namespace Microsoft.DotNet.Interactive.Http
+namespace Microsoft.DotNet.Interactive.AspNetCore
 {
-    public static class EndpointRouteBuilderExtensions
+    public static class InteractiveEndpointRouteBuilderExtensions
     {
         private static int __AspNet_NextEndpointOrder;
 
-        public static IEndpointConventionBuilder MapAction(
+        public static IEndpointConventionBuilder MapInteractive(
             this IEndpointRouteBuilder endpoints,
             string pattern,
             RequestDelegate requestDelegate)
         {
             var order = __AspNet_NextEndpointOrder--;
-            var builder = endpoints.MapGet(pattern, requestDelegate);
+            var builder = endpoints.Map(pattern, requestDelegate);
             builder.Add(b => ((RouteEndpointBuilder)b).Order = order);
             return builder;
         }
