+++ src/Microsoft.DotNet.Interactive.CSharp/CSharpKernel.cs
@@ -9,9 +9,6 @@ using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Diagnostics;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Completion;
 using Microsoft.CodeAnalysis.CSharp;
@@ -23,7 +20,6 @@ using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.Extensions;
 using Microsoft.DotNet.Interactive.Formatting;
 using Microsoft.DotNet.Interactive.Utility;
-using Microsoft.Extensions.Hosting;
 using XPlot.Plotly;
 
 using CompletionItem = Microsoft.DotNet.Interactive.Events.CompletionItem;
@@ -60,15 +56,8 @@ namespace Microsoft.DotNet.Interactive.CSharp
                              "System.Text",
                              "System.Collections",
                              "System.Collections.Generic",
-                             "System.Threading",
                              "System.Threading.Tasks",
-                             "System.Linq",
-                             "System.Net.Http",
-                             "Microsoft.AspNetCore",
-                             "Microsoft.AspNetCore.Builder",
-                             "Microsoft.AspNetCore.Http",
-                             "Microsoft.AspNetCore.Routing",
-                             "Microsoft.Extensions.Logging")
+                             "System.Linq")
                          .AddReferences(
                              typeof(Enumerable).Assembly,
                              typeof(IEnumerable<>).Assembly,
@@ -76,11 +65,7 @@ namespace Microsoft.DotNet.Interactive.CSharp
                              typeof(Kernel).Assembly,
                              typeof(CSharpKernel).Assembly,
                              typeof(PocketView).Assembly,
-                             typeof(PlotlyChart).Assembly,
-                             typeof(Host).Assembly,
-                             typeof(WebHost).Assembly,
-                             typeof(Controller).Assembly,
-                             typeof(DeveloperExceptionPageMiddleware).Assembly);
+                             typeof(PlotlyChart).Assembly);
 
         private readonly AssemblyBasedExtensionLoader _extensionLoader = new AssemblyBasedExtensionLoader();
         private string _currentDirectory;
