+++ src/Microsoft.DotNet.Interactive.Http/Microsoft.DotNet.Interactive.Http.csproj
@@ -16,9 +16,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.14.0" />
-    <PackageReference Include="PocketLogger" Version="0.4.0">
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
   </ItemGroup>
 
   <ItemGroup>
