+++ src/dotnet-interactive/Program.cs
@@ -11,6 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.DotNet.Interactive.App.CommandLine;
+using Microsoft.DotNet.Interactive.AspNetCore;
 using Microsoft.DotNet.Interactive.Http;
 using Microsoft.DotNet.Interactive.Jupyter;
 using Microsoft.Extensions.DependencyInjection;
@@ -36,7 +37,7 @@ namespace Microsoft.DotNet.Interactive.App
             typeof(Startup).Assembly, // dotnet-interactive.dll
             typeof(Kernel).Assembly, // Microsoft.DotNet.Interactive.dll
             typeof(Shell).Assembly, // Microsoft.DotNet.Interactive.Jupyter.dll
-            typeof(AspNetKernelExtensions).Assembly, // Microsoft.DotNet.Interactive.Http.dll
+            typeof(AspNetCoreKernelExtensions).Assembly, // Microsoft.DotNet.Interactive.AspNetCore.dll
         };
 
         internal static IDisposable StartToolLogging(StartupOptions options)
